Remove useMediaReducer folder and move reducer to media/local/reducer.js.
