+++ assets/src/edit-story/app/media/local/reducer.js
@@ -19,7 +19,7 @@
 /**
  * Internal dependencies
  */
-import * as types from './types';
+import * as types from '../types';
 
 export const INITIAL_STATE = {
   media: [],
